refactor: :art: changing component structure
