+++ src/components/profile/profile.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import profileImage from '../../../assets/profile.png';
+import profileImage from '../../assets/profile.png';
 
 function Profile() {
   return (
