+++ src/pages/HomePage/HomePage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../../components/organisms/header/header';
+import Header from '../../components/header/header';
 
 function HomePage() {
   return (
